services/appraisal-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.8.2
flask-migrate==4.0.5